+++ src/main/java/mod/gottsch/forge/treasure2/datagen/Recipes.java
@@ -176,6 +176,209 @@ public class Recipes extends RecipeProvider {
 					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
 					.save(recipe);
 
+			// gold skull chest
+			ShapedRecipeBuilder.shaped(RecipeCategory.MISC, TreasureBlocks.GOLD_SKULL_CHEST.get())
+					.pattern("ggt")
+					.pattern("gxg")
+					.pattern("ggg")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('x', Items.SKELETON_SKULL)
+					.define('g', Items.GOLD_INGOT)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			// crystal skull chest
+			ShapedRecipeBuilder.shaped(RecipeCategory.MISC, TreasureBlocks.CRYSTAL_SKULL_CHEST.get())
+					.pattern("ggt")
+					.pattern("gxg")
+					.pattern("ggg")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('x', Items.SKELETON_SKULL)
+					.define('g', Items.DIAMOND)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			// wither planks
+			ShapelessRecipeBuilder.shapeless(RecipeCategory.MISC, TreasureBlocks.WITHER_PLANKS.get())
+					.requires(TreasureBlocks.WITHER_LOG.get())
+					.unlockedBy("has_log", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureBlocks.WITHER_LOG.get()))
+					.save(recipe);
+
+
+			// gravestones
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.SKULL_CROSSBONES.get())
+					.pattern("  t")
+					.pattern(" b ")
+					.pattern("b b")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('b', Items.BONE)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE1_COBBLESTONE.get())
+					.pattern("  t")
+					.pattern("   ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.COBBLESTONE)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE1_MOSSY_COBBLESTONE.get())
+					.pattern("  t")
+					.pattern("   ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.MOSSY_COBBLESTONE)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE1_OBSIDIAN.get())
+					.pattern("  t")
+					.pattern("   ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.OBSIDIAN)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE1_POLISHED_GRANITE.get())
+					.pattern("  t")
+					.pattern("   ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.POLISHED_GRANITE)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE1_SMOOTH_QUARTZ.get())
+					.pattern("  t")
+					.pattern("   ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.SMOOTH_QUARTZ)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE1_STONE.get())
+					.pattern("  t")
+					.pattern("   ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.STONE)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			// gravestone 2
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE2_COBBLESTONE.get())
+					.pattern("  t")
+					.pattern(" s ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.COBBLESTONE)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE2_MOSSY_COBBLESTONE.get())
+					.pattern("  t")
+					.pattern(" s ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.MOSSY_COBBLESTONE)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE2_OBSIDIAN.get())
+					.pattern("  t")
+					.pattern(" s ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.OBSIDIAN)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE2_POLISHED_GRANITE.get())
+					.pattern("  t")
+					.pattern(" s ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.POLISHED_GRANITE)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE2_SMOOTH_QUARTZ.get())
+					.pattern("  t")
+					.pattern(" s ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.SMOOTH_QUARTZ)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE2_STONE.get())
+					.pattern("  t")
+					.pattern(" s ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.STONE)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			// gravestone 3
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE3_COBBLESTONE.get())
+					.pattern(" st")
+					.pattern(" s ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.COBBLESTONE)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE3_MOSSY_COBBLESTONE.get())
+					.pattern(" st")
+					.pattern(" s ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.MOSSY_COBBLESTONE)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE3_OBSIDIAN.get())
+					.pattern(" st")
+					.pattern(" s ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.OBSIDIAN)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE3_POLISHED_GRANITE.get())
+					.pattern(" st")
+					.pattern(" s ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.POLISHED_GRANITE)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE3_SMOOTH_QUARTZ.get())
+					.pattern(" st")
+					.pattern(" s ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.SMOOTH_QUARTZ)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
+			ShapedRecipeBuilder.shaped(RecipeCategory.DECORATIONS, TreasureBlocks.GRAVESTONE3_STONE.get())
+					.pattern(" st")
+					.pattern(" s ")
+					.pattern(" s ")
+					.define('t', TreasureItems.TREASURE_TOOL.get())
+					.define('s', Items.STONE)
+					.unlockedBy("has_tool", InventoryChangeTrigger.TriggerInstance.hasItems(TreasureItems.TREASURE_TOOL.get()))
+					.save(recipe);
+
 	        // copper weapons smelting
 			SimpleCookingRecipeBuilder.smelting(Ingredient.of(TreasureItems.CHIPPED_COPPER_SHORT_SWORD.get()), RecipeCategory.COMBAT, 
 	                Items.COPPER_INGOT, 1.0f, 200)